requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
Django==4.2.5
django-cors-headers==4.2.0
django-import-export==3.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
